added feature request #5444. Also added to scripts
git-svn-id: http://sextante.googlecode.com/svn/trunk/soft/bindings/qgis-plugin@140 881b9c09-3ef8-f3c2-ec3d-21d735c97f4d

+++ src/sextante/modeler/DeleteModelAction.py
@@ -1,6 +1,7 @@
 from sextante.gui.ContextAction import ContextAction
 from sextante.modeler.ModelerAlgorithm import ModelerAlgorithm
 import os
+from PyQt4 import QtGui
 
 class DeleteModelAction(ContextAction):
 
@@ -11,5 +12,9 @@ class DeleteModelAction(ContextAction):
         return isinstance(self.alg, ModelerAlgorithm)
 
     def execute(self):
-        os.remove(self.alg.descriptionFile)
-        self.toolbox.updateTree()
+        reply = QtGui.QMessageBox.question(None, 'Confirmation',
+                            "Are you sure you want to delete this model?", QtGui.QMessageBox.Yes |
+                            QtGui.QMessageBox.No, QtGui.QMessageBox.No)
+        if reply == QtGui.QMessageBox.Yes:
+            os.remove(self.alg.descriptionFile)
+            self.toolbox.updateTree()
